Implement the changeGreeting: method
Get the text from the text field and set the userName propriety (model object) to its value. Create a new NSString from the model object and test if its length is 0. If it is then put “World” as default NSString. Finally construct the greeting and set the label text to its value.